backend/requirements.txt
# SE Insight Backend Dependencies

# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket support
websockets==12.0

# CORS and multipart support
python-multipart==0.0.6

# Audio processing and ASR
numpy==1.24.3
SpeechRecognition==3.10.0
pyaudio==0.2.11  # For microphone input (optional)

# Advanced ASR (optional but recommended)
openai-whisper==20231117
torch==2.1.0  # Required for Whisper

# SE terminology extraction
keybert==0.8.3
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Future enhancements (optional)
# neo4j==5.14.1  # For advanced knowledge base
# openai==1.3.5  # For LLM integration
# redis==5.0.1   # For caching and session management